logging-service/pyproject.toml
[project]
name = "chrona-logging-service"
version = "0.1.0"
description = "A logging service exclusive for Chrona."
authors = [
    {name = "[Miracle Mayowa  Adebunmi",email = "blueprime91@gmail.com"}
]
license = {text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "python-decouple (>=3.8,<4.0)",
    "redis (>=5.2.1,<6.0.0)",
    "sentry-sdk (>=2.25.1,<3.0.0)",
    "prometheus-client (>=0.21.1,<0.22.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "logdna (>=1.18.12,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
